@@ -2,23 +2,56 @@
 
 declare(strict_types=1);
 
-return [
-    /**
-     * Default database connection
-     */
-    'default'   => 'default',
+use Cycle\Database\Config;
 
-    'databases' => [
-        'default' => [
-            'driver' => 'sqlite'
+return [
+    'logger' => [
+        'default' => null,
+        'drivers' => [
+            'runtime' => 'stdout'
         ],
     ],
 
-    'drivers'   => [
-        'sqlite' => [
-            'driver'     => \Spiral\Database\Driver\SQLite\SQLiteDriver::class,
-            'connection' => 'sqlite:' . directory('root') . 'app.db',
-            'profiling'  => true,
+    /**
+     * Default database connection
+     */
+    'default' => 'default',
+
+    /**
+     * The Spiral/Database module provides support to manage multiple databases
+     * in one application, use read/write connections and logically separate
+     * multiple databases within one connection using prefixes.
+     *
+     * To register a new database simply add a new one into
+     * "databases" section below.
+     */
+    'databases' => [
+        'default' => [
+            'driver' => 'mariadb',
         ],
-    ]
+    ],
+
+    /**
+     * Each database instance must have an associated connection object.
+     * Connections used to provide low-level functionality and wrap different
+     * database drivers. To register a new connection you have to specify
+     * the driver class and its connection options.
+     */
+    'drivers' => [
+        'runtime' => new Config\SQLiteDriverConfig(
+            connection: new Config\SQLite\MemoryConnectionConfig(),
+            queryCache: true
+        ),
+        'persistent' => new Config\SQLiteDriverConfig(
+            connection: new Config\SQLite\FileConnectionConfig(directory('root') . 'app.db'),
+            queryCache: true
+        ),
+        'mariadb' => new Config\MySQLDriverConfig(
+            connection: new Config\MySQL\DsnConnectionConfig(
+                dsn: 'mysql:host=mariadb;dbname=arena',
+                user: 'arena-user',
+                password: 'arena-password'
+            )
+        ),
+    ],
 ];
